+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Fix display of languages in the terminal.
 - Show message if not specified --speak-source or --speak-translation options in --audio-only mode.
+- Change the CLI option "translator" to "locale".
 
 ## [1.0.3] - 2018-07-04
 ### Added